requirements.txt
coverage
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-yasg
flake8
psycopg2-binary
python-dotenv